backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
crewai[google-genai]>=0.80.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pdfminer.six>=20221105
